+++ src/components/EntriesTest/NewEntryForm/NewEntryForm.component.tsx
