+++ src/components/EntriesTest/NewEntryForm/styled.tsx
